Add options for changing text color.
