+++ src/Widgets/NestedTimerWindow.h
@@ -7,13 +7,13 @@
 #include "DraggableQWidget.h"
 #include "UpdatableGameWindow.h"
 #include "ScaleableLabel.h"
-#include "BackgroundChangerWidgetHelper.h"
+#include "ColorChangerWidgetHelper.h"
 
 #include <QLCDNumber>
 #include <QMenu>
 #include <QGridLayout>
 
-class NestedTimerWindow: public DraggableQWidget, public UpdatableGameWindow, public BackgroundChangerWidgetHelper
+class NestedTimerWindow: public DraggableQWidget, public UpdatableGameWindow, public ColorChangerWidgetHelper
 {
 public:
     NestedTimerWindow();
