+++ src/Widgets/SimpleTimerWindow.cpp
@@ -26,7 +26,7 @@ SimpleTimerWindow::SimpleTimerWindow(bool showControls)
     setWindowTitle("Timer");
     setWindowFlags(Qt::Window | Qt::NoDropShadowWindowHint | Qt::FramelessWindowHint | Qt::WindowStaysOnTopHint);
 
-    SetupBackgroundChanger(this);
+    SetupColorChanger(this);
 
     resizeBorder = 4;
 
@@ -37,7 +37,7 @@ SimpleTimerWindow::SimpleTimerWindow(bool showControls)
     });
 
     contextMenu = new QMenu(this);
-    AddBackgroundOptionsToMenu(contextMenu);
+    AddColorOptionsToMenu(contextMenu, true, true);
     contextMenu->addSeparator();
     contextMenu->addAction(actionExit);
